refactor(user): split error classes and use withErrorHandling + jsonResponse in routes
