+++ src/domains/user/errors/index.ts
@@ -5,10 +5,16 @@
  * @remarks
  * Surfaces typed domain, validation, and authorization errors raised during
  * user profile lookups together with factory helpers that construct them.
- * @see {@link module:domains/user/errors/user-not-found-error} for error class definitions
+ * Each error class and its factory live in the same file (one error per file),
+ * mirroring the convention used in the anime and music domains.
+ *
+ * @see {@link module:domains/user/errors/user-not-found-error} — 404 not found
+ * @see {@link module:domains/user/errors/user-invalid-id-error} — 400 bad id
+ * @see {@link module:domains/user/errors/user-unauthorized-error} — 400 unauthorized
+ * @see {@link module:domains/user/errors/user-profile-conflict-error} — 409 conflict
  * @example
  * ```typescript
- * import { userNotFound, UserUnauthorizedError } from '@user/errors'
+ * import { userNotFound, UserProfileConflictError } from '@user/errors'
  *
  * throw userNotFound('user-123')
  * ```
@@ -16,13 +22,17 @@
 
 export {
   UserNotFoundError,
-  UserInvalidIdError,
-  UserUnauthorizedError,
-  UserProfileConflictError,
-} from './user-error-classes'
-export {
   userNotFound,
-  userInvalidId,
-  userUnauthorized,
-  userProfileConflict,
 } from './user-not-found-error'
+export {
+  UserInvalidIdError,
+  userInvalidId,
+} from './user-invalid-id-error'
+export {
+  UserUnauthorizedError,
+  userUnauthorized,
+} from './user-unauthorized-error'
+export {
+  UserProfileConflictError,
+  userProfileConflict,
+} from './user-profile-conflict-error'
