+++ src/domains/user/errors/user-not-found-error.ts
@@ -1,92 +1,51 @@
 /**
- * Factory helpers for user lookup and authorization domain errors.
+ * Domain error raised when no user profile exists for the given identifier.
  *
  * @module domains/user/errors/user-not-found-error
  * @remarks
- * Factories return error instances so callers can `throw` them from services or route handlers.
- * The error classes themselves live in {@link user-error-classes} and are re-exported here so
- * existing `@user/errors/user-not-found-error` imports keep working.
+ * Thrown by {@link userService.getUserProfile}, {@link userService.updateUserProfile}
+ * (after `updateProfile` returns `undefined`) and any code that resolves a profile row
+ * by id. Surfaces as HTTP **404** via {@link mapErrorToHttp} (`USER_NOT_FOUND` is in the
+ * not-found set).
+ *
+ * Distinct from {@link UserInvalidIdError}, which covers malformed route or query
+ * parameters and maps to 400.
+ *
+ * @see {@link mapErrorToHttp}
  */
-import {
-  UserInvalidIdError,
-  UserNotFoundError,
-  UserProfileConflictError,
-  UserUnauthorizedError,
-} from '@user/errors/user-error-classes'
-
-// UserNotFoundError, UserInvalidIdError, UserUnauthorizedError, and
-// UserProfileConflictError are re-exported via the barrel at `@user/errors`.
-// Import them from there or from `@user/errors/user-error-classes` directly.
+import { DomainError } from '@shared/errors/app-error'
+import { ErrorCodes } from '@shared/errors/codes'
 
 /**
- * Creates a {@link UserNotFoundError} for the given user identifier.
+ * Error thrown when no user profile exists for the given identifier.
+ *
+ * @remarks
+ * - **Code:** `USER_NOT_FOUND`
+ * - **Details:** `{ id: string }`
+ * - **HTTP:** 404 (`DomainError` + `NOT_FOUND_DOMAIN_CODES` in {@link mapErrorToHttp})
+ *
+ * @example
+ * ```typescript
+ * if (!row) throw userNotFound(targetId)
+ * ```
+ */
+export class UserNotFoundError extends DomainError {
+  constructor(id: string) {
+    super(ErrorCodes.USER_NOT_FOUND, 'User not found', { id })
+  }
+}
+
+/**
+ * Factory for {@link UserNotFoundError}.
  *
  * @param id - User identifier that was not found
  * @returns A {@link UserNotFoundError} ready to be thrown
- * @throws Does not throw; returns an error instance for the caller to throw
- * @remarks
- * Prefer this factory over `new UserNotFoundError()` in services to keep
- * construction consistent and reduce import surface area.
  * @see {@link UserNotFoundError}
  * @example
  * ```typescript
- * throw userNotFound(targetId)
+ * throw userNotFound('user-123')
  * ```
  */
 export const userNotFound = (id: string) => {
   return new UserNotFoundError(id)
-}
-
-/**
- * Creates a {@link UserInvalidIdError} for an invalid identifier.
- *
- * @param rawId - Unparsed identifier from the request
- * @returns A {@link UserInvalidIdError} ready to be thrown
- * @throws Does not throw; returns an error instance for the caller to throw
- * @remarks
- * Use at validation boundaries when Zod or manual parsing rejects a user ID.
- * @see {@link UserInvalidIdError}
- * @example
- * ```typescript
- * throw userInvalidId(params.userId)
- * ```
- */
-export const userInvalidId = (rawId: unknown) => {
-  return new UserInvalidIdError(rawId)
-}
-
-/**
- * Creates a {@link UserUnauthorizedError} for a denied access attempt.
- *
- * @param userId - User identifier that was denied access
- * @returns A {@link UserUnauthorizedError} ready to be thrown
- * @throws Does not throw; returns an error instance for the caller to throw
- * @remarks
- * Typically thrown after a {@link userPolicies} check returns `false`.
- * @see {@link UserUnauthorizedError}
- * @see {@link userPolicies}
- * @example
- * ```typescript
- * throw userUnauthorized(targetId)
- * ```
- */
-export const userUnauthorized = (userId: string) => {
-  return new UserUnauthorizedError(userId)
-}
-
-/**
- * Creates a {@link UserProfileConflictError} for a duplicate profile write.
- *
- * @param userId - User identifier for which a profile already exists
- * @returns A {@link UserProfileConflictError} ready to be thrown
- * @remarks
- * Maps to HTTP 409 via {@link mapErrorToHttp}.
- * @see {@link UserProfileConflictError}
- * @example
- * ```typescript
- * if (existing) throw userProfileConflict(userId)
- * ```
- */
-export const userProfileConflict = (userId: string) => {
-  return new UserProfileConflictError(userId)
 }
